feat(engine): extract isFailingCheckSummary to gittensory-engine (#4256)
