+++ packages/gittensory-engine/src/reward-risk.ts
@@ -3,14 +3,15 @@
 //
 // Unlike the earlier self-contained extractions, reward-risk sits on top of the maintainer signal stack in
 // `src/signals/engine.ts` (`buildRoleContext`, `buildLaneAdvice`, `buildCollisionReport`, `buildQueueHealth`,
-// `buildRepoFitRecommendation`, `buildContributorIntakeHealth`, `buildPullRequestReviewIntelligence`) and on
-// `isFailingCheckSummary` from `src/signals/local-branch.ts`. Those builders are not yet extracted (and are
-// far too large to port under the per-issue size cap), so — rather than reach back into `src/`, which this
-// package must never do — they are DEPENDENCY-INJECTED via `RewardRiskEngineDeps`. The
-// `src/signals/reward-risk.ts` shim binds the real `src` builders; a follow-up issue can drop the injection
-// once they have engine homes.
+// `buildRepoFitRecommendation`, `buildContributorIntakeHealth`, `buildPullRequestReviewIntelligence`) from
+// `src/signals/engine.ts`. Those builders are not yet extracted (and are far too large to port under the
+// per-issue size cap), so — rather than reach back into `src/`, which this package must never do — they are
+// DEPENDENCY-INJECTED via `RewardRiskEngineDeps`. `isFailingCheckSummary` now lives in
+// `./signals/check-summary.js` (#4256). The `src/signals/reward-risk.ts` shim binds the real `src` builders;
+// a follow-up issue can drop the injection once they have engine homes.
 import type { ScorePreviewResult } from "./scoring/preview.js";
 import { buildScorePreview } from "./scoring/preview.js";
+import { isFailingCheckSummary } from "./signals/check-summary.js";
 import { nowIso } from "./utils/json.js";
 import type {
   CheckSummaryRecord,
@@ -60,10 +61,10 @@ export type PullRequestReviewabilityInput = {
 };
 
 /**
- * The `src/signals/engine.ts` + `src/signals/local-branch.ts` builders reward-risk depends on, injected so
- * this package stays free of any `src/` import. The real `src`-typed builders bind cleanly: their argument
- * records are wider than (assignable from) these engine mirrors, and their richer return types are
- * covariantly assignable to the narrowed views above.
+ * The `src/signals/engine.ts` builders reward-risk depends on, injected so this package stays free of any
+ * `src/` import. The real `src`-typed builders bind cleanly: their argument records are wider than (assignable
+ * from) these engine mirrors, and their richer return types are covariantly assignable to the narrowed views
+ * above. `isFailingCheckSummary` is imported directly from `./signals/check-summary.js` (#4256).
  */
 export type RewardRiskEngineDeps = {
   buildRoleContext: (args: {
@@ -104,7 +105,6 @@ export type RewardRiskEngineDeps = {
     collisions: CollisionReport,
   ) => { level: "healthy" | "watch" | "strained" | "blocked" };
   buildPullRequestReviewIntelligence: (args: PullRequestReviewabilityInput) => PullRequestReviewIntelligenceView;
-  isFailingCheckSummary: (check: CheckSummaryRecord) => boolean;
 };
 
 export type RewardRiskActionKind =
@@ -541,7 +541,7 @@ export function buildMaintainerNoiseReport(
 export function buildPullRequestReviewability(args: PullRequestReviewabilityInput, deps: RewardRiskEngineDeps): PullRequestReviewability {
   const intelligence = deps.buildPullRequestReviewIntelligence(args);
   const pr = args.pullRequest;
-  const failingChecks = args.checks.filter(deps.isFailingCheckSummary).length;
+  const failingChecks = args.checks.filter(isFailingCheckSummary).length;
   const broadDiff = intelligence.changeSummary.fileCount >= 12 || intelligence.changeSummary.additions + intelligence.changeSummary.deletions >= 800;
   const noiseSources = [
     ...(pr?.state && pr.state !== "open" ? [`PR is ${pr.state}.`] : []),
